Introduce ConfigurationServiceHelper
The ConfigurationServiceHelper is a small singleton used to add easily set ConfigurationService stores.

It allows for the injection of multiple initializer handlers and updater handlers.
Initializer handlers are called when the configuration is first fetched.
Updater handlers are called when the configuration is updated.

It includes a log4j2 initializer and updater. This handler looks for configuration URIs under the config property "log4j.config.uri".

Issue #9